Moved init functionality back to kernel from loader64. Made tests work
