--- alkos/kernel/arch/x86_64/common-loader-64-kernel/panic.nasm
@@ -1,6 +1,0 @@
-extern KernelPanic
-
-%macro panic 1
-    mov rdi, %1
-    call KernelPanic
-%endmacro
