--- alkos/kernel/arch/x86_64/common-loader-64-kernel/trace.nasm
@@ -1,20 +1,0 @@
-extern QemuTerminalWriteString
-
-%ifdef __USE_DEBUG_OUTPUT__
-
-%macro trace 1
-    push rdi
-    mov rdi, %1
-    call QemuTerminalWriteString
-    pop rdi
-%endmacro
-
-%else
-
-%macro trace 1
-%endmacro
-
-%endif
-
-
-
