+++ alkos/kernel/arch/x86_64/kernel/abi/panic.cpp
@@ -9,6 +9,7 @@
 extern "C" void NO_RET KernelPanic(const char *msg)
 {
     TerminalWriteError("[ KERNEL PANIC ]\n");
+    TerminalWriteError(" Kernel Panic Kernel ");
     TerminalWriteError(msg);
     TerminalWriteError("\n");
