+++ alkos/kernel/arch/x86_64/kernel/arch_utils.cpp
