+++ alkos/kernel/src/init.cpp
@@ -1,4 +1,4 @@
-#include <init.hpp>
+#include <loader64_main.hpp>
 
 /* Internal includes */
 #include <libssp.h>
