+++ scripts/tests/runner.py
@@ -42,6 +42,7 @@ TEST_RUNS: list[TestRun] = [
         [
             "SnprintfTest_BufferSizeHandling",  # Fails <- issue added on low prio
             "SnprintfTest_EdgeCases",  # Fails <- issue added on low prio
+            # "TestGroupBase_SimpleExceptionTest"
         ]
     )
 ]
